Begin refactoring code for computing site metrics
