+++ frontend/src/components/Forms/index.js
@@ -3,4 +3,5 @@ export * from './DownloadButton'
 export * from './DropZone'
 export * from './ProposalDetails'
 export * from './StudiesDownload'
+export * from './SiteMetricsDownload'
 export * from './TaskManager'
