+++ frontend/src/components/Tables/SitesTable.js
@@ -1,14 +1,9 @@
 import React, { useContext, useEffect } from 'react'
 import MaterialTable from 'material-table'
 import { StoreContext } from '../../contexts/StoreContext'
-import { SiteDetailPanel, dayCount, displayRatio } from './DetailPanels'
+import { SiteDetailPanel } from './DetailPanels'
 import { EnrollmentBar } from '../Widgets/EnrollmentBar'
-
-const invalidDisplay = 'N/A'
-
-const ratioAsWholeNumberString = (a, b) => {
-    return b === 0 ? invalidDisplay : Math.round(a / b)
-}
+import { computeMetrics } from '../../utils/sites'
 
 export const SitesTable = props => {
     let { title, sites } = props
@@ -26,16 +21,7 @@ export const SitesTable = props => {
                      site.protocol = shortTitle
                 }
 
-                site.protocolToFpfv = dayCount(site.dateRegPacketSent, site.fpfv)
-                site.contractExecutionTime = dayCount(site.dateContractSent, site.dateContractExecution)
-                site.sirbApprovalTime = dayCount(site.dateIrbSubmission, site.dateIrbApproval)
-                site.siteOpenToFpfv = dayCount(site.dateSiteActivated, site.fpfv)
-                site.protocolToLpfv = dayCount(site.dateSiteActivated, site.lpfv)
-                site.percentConsentedPtsRandomized = displayRatio(site.patientsEnrolledCount, site.patientsConsentedCount)
-                site.actualToExpectedRandomizedPtRatio = displayRatio(site.patientsEnrolledCount, site.patientsExpectedCount)
-                site.ratioRandomizedPtsDropout = displayRatio(site.patientsWithdrawnCount, site.patientsEnrolledCount)
-                site.majorProtocolDeviationsPerRandomizedPt = displayRatio( site.protocolDeviationsCount, site.patientsEnrolledCount)
-                site.queriesPerConsentedPatient = ratioAsWholeNumberString(site.queriesCount, site.patientsConsentedCount )
+                computeMetrics(site)                
                 site.shortDescription = site.protocol.shortDescription
 
                 // Enrollment
