+++ frontend/src/utils/sites/index.js
@@ -1,1 +1,6 @@
+export * from './computeMetrics'
+export * from './convertEnrollmentData'
+export * from './dayCount'
+export * from './displayRatio'
+export * from './invalidDisplay'
 export * from './isSiteActive'
