+++ frontend/src/utils/sites/invalidDisplay.js
@@ -1,0 +1,2 @@
+export const invalidDisplay = 'N/A'
+
