+++ frontend/src/views/Studies/List.js
@@ -4,7 +4,7 @@ import api from '../../Api'
 import { Title, Subsubheading, Paragraph } from '../../components/Typography'
 import { StudiesTable } from '../../components/Tables'
 import { Grid, List, ListItem, ListItemText } from '@material-ui/core'
-import { DropZone, DownloadButton as TemplateDownload, StudiesDownloadForm } from '../../components/Forms'
+import { DropZone, DownloadButton as TemplateDownload, StudiesDownloadForm, SiteMetricsDownload } from '../../components/Forms'
 import { DataUploadHelper } from '../../components/Helper'
 
 export const StudiesListPage = (props) => {
@@ -28,6 +28,7 @@ export const StudiesListPage = (props) => {
           <TemplateDownload path={api.download('study-profile')} tooltip="Download Study Profile CSV Template" />
           <DropZone endpoint={api.uploadStudyProfile} method="POST" />
           <StudiesDownloadForm />
+          <SiteMetricsDownload />
         </Grid>
       </Grid>
